feat(qdrant-rag)!: extract QdrantRag + QdrantRagProvider into @mcp-abap-adt/qdrant-rag
Extract vector RAG implementation (QdrantRag) and its provider (QdrantRagProvider) from core into a new @mcp-abap-adt/qdrant-rag package. This completes the separation of concerns by isolating external vector store dependencies (Qdrant API calls via fetch) from the core agent library.

Changes:
- Create packages/qdrant-rag with QdrantRag, QdrantRagProvider, and tests
- Remove axios from @mcp-abap-adt/llm-agent dependencies (no longer used by core)
- Add back-compat re-exports in llm-agent-server for QdrantRag and QdrantRagProvider
- Update server to import QdrantRag from @mcp-abap-adt/qdrant-rag
- Include qdrant-rag in root build/clean scripts

All 15 qdrant-rag tests pass. Core library now 100% backend-agnostic.
