+++ package.json
@@ -5,8 +5,8 @@
   "workspaces": ["packages/*"],
   "scripts": {
     "prebuild": "npm run --workspace @mcp-abap-adt/llm-agent-server prebuild",
-    "build": "tsc -b packages/llm-agent packages/openai-llm packages/anthropic-llm packages/openai-embedder packages/ollama-embedder packages/sap-aicore-llm packages/sap-aicore-embedder packages/deepseek-llm packages/llm-agent-server",
-    "clean": "tsc -b --clean packages/llm-agent packages/openai-llm packages/anthropic-llm packages/openai-embedder packages/ollama-embedder packages/sap-aicore-llm packages/sap-aicore-embedder packages/deepseek-llm packages/llm-agent-server",
+    "build": "tsc -b packages/llm-agent packages/openai-llm packages/anthropic-llm packages/openai-embedder packages/ollama-embedder packages/qdrant-rag packages/sap-aicore-llm packages/sap-aicore-embedder packages/deepseek-llm packages/llm-agent-server",
+    "clean": "tsc -b --clean packages/llm-agent packages/openai-llm packages/anthropic-llm packages/openai-embedder packages/ollama-embedder packages/qdrant-rag packages/sap-aicore-llm packages/sap-aicore-embedder packages/deepseek-llm packages/llm-agent-server",
     "lint": "biome check --write packages",
     "lint:check": "biome check packages",
     "format": "biome format --write packages",
