+++ packages/llm-agent-server/package.json
@@ -65,6 +65,7 @@
   "devDependencies": {
     "@mcp-abap-adt/ollama-embedder": "*",
     "@mcp-abap-adt/openai-embedder": "*",
+    "@mcp-abap-adt/qdrant-rag": "*",
     "@mcp-abap-adt/sap-aicore-embedder": "*",
     "@opentelemetry/api": "^1.9.1",
     "@sap-ai-sdk/ai-api": "^2.0.0",
