+++ packages/llm-agent-server/src/smart-agent/providers.ts
@@ -20,11 +20,11 @@ import type {
 import {
   builtInEmbedderFactories,
   InMemoryRag,
-  QdrantRag,
   VectorRag,
 } from '@mcp-abap-adt/llm-agent';
 import { OllamaRag } from '@mcp-abap-adt/ollama-embedder';
 import { OpenAIProvider } from '@mcp-abap-adt/openai-llm';
+import { QdrantRag } from '@mcp-abap-adt/qdrant-rag';
 import {
   type SapAICoreCredentials,
   SapCoreAIProvider,
