+++ packages/llm-agent/src/rag/index.ts
@@ -14,8 +14,6 @@ export {
   TranslatePreprocessor,
 } from './preprocessor.js';
 export * from './providers/index.js';
-export type { QdrantRagConfig } from './qdrant-rag.js';
-export { QdrantRag } from './qdrant-rag.js';
 export {
   FallbackQueryEmbedding,
   QueryEmbedding,
