+++ packages/qdrant-rag/README.md
@@ -1,0 +1,5 @@
+# @mcp-abap-adt/qdrant-rag
+
+QdrantRag vector store and QdrantRagProvider for @mcp-abap-adt/llm-agent.
+
+Provides vector search capabilities using Qdrant as the backend.
