+++ packages/qdrant-rag/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@mcp-abap-adt/llm-agent",
-  "version": "10.0.0",
-  "description": "Core interfaces, types, and lightweight default implementations for LLM agent orchestration.",
+  "name": "@mcp-abap-adt/qdrant-rag",
+  "version": "11.0.0",
+  "description": "QdrantRag vector store and QdrantRagProvider for @mcp-abap-adt/llm-agent.",
   "type": "module",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -23,8 +23,7 @@
     "test": "node --import tsx/esm --test --test-reporter=spec 'src/**/*.test.ts'"
   },
   "dependencies": {
-    "@sap-ai-sdk/orchestration": "^2.9.0",
-    "zod": "^3.25.0"
+    "@mcp-abap-adt/llm-agent": "*"
   },
   "license": "MIT",
   "repository": {
@@ -33,8 +32,5 @@
   },
   "publishConfig": {
     "access": "public"
-  },
-  "devDependencies": {
-    "@types/node": "^25.6.0"
   }
 }
