+++ packages/qdrant-rag/src/index.ts
@@ -1,0 +1,4 @@
+export type { QdrantRagConfig } from './qdrant-rag.js';
+export { QdrantRag } from './qdrant-rag.js';
+export type { QdrantRagProviderConfig } from './qdrant-rag-provider.js';
+export { QdrantRagProvider } from './qdrant-rag-provider.js';
